Add an exclude_patterns configuration parameter
Using exclude_patterns in your configuration file, you can exclude log
lines that match one or more regexes.  This is useful if you have log
lines that you cannot easily prevent from showing up in your logs and
that do not want to be sent to the remote syslog receiver.

The excludes_patterns is global to all log files and not tied to
specific files. So make sure that your regex is specific enough to not
catch valid lines unintentionally.